did refactor hm -> hw
