+++ log.0.0.txt
@@ -1,28 +1,28 @@
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [38, 3, 36, 16, 15, 4, 9, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 38, 36, 16, 15, 4, 9, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 38, 36, 16, 15, 4, 9, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 38, 36, 16, 15, 9, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 38, 36, 16, 15, 9, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 38, 36, 16, 15, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 38, 36, 16, 15, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 38, 36, 16, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 38, 36, 16, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 16, 38, 36, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 16, 38, 36, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 16, 36, 38, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 16, 36, 38, 46]
-Mar 31, 2023 3:17:25 PM seminar2.hm.Task1 BubbleSort
+Mar 31, 2023 3:17:25 PM seminar2.hw.Task1 BubbleSort
 INFO: [3, 4, 9, 15, 16, 36, 38, 46]
