+++ src/seminar2/hw/View.java
@@ -1,4 +1,4 @@
-package seminar2.hm;
+package seminar2.hw;
 
 import java.util.Arrays;
